# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PreClinik API — e-learning platform for Algerian medical students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules/{moduleId}/lessons/{lessonId}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List Q-bank questions for a lesson
      parameters:
        - $ref: "#/components/parameters/moduleId"
        - $ref: "#/components/parameters/lessonId"
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a Q-bank question (admin)
      parameters:
        - $ref: "#/components/parameters/moduleId"
        - $ref: "#/components/parameters/lessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /modules/{moduleId}/lessons/{lessonId}/questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a Q-bank question (admin)
      parameters:
        - $ref: "#/components/parameters/moduleId"
        - $ref: "#/components/parameters/lessonId"
        - $ref: "#/components/parameters/questionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a Q-bank question (admin)
      parameters:
        - $ref: "#/components/parameters/moduleId"
        - $ref: "#/components/parameters/lessonId"
        - $ref: "#/components/parameters/questionId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  parameters:
    moduleId:
      name: moduleId
      in: path
      required: true
      schema:
        type: string
    lessonId:
      name: lessonId
      in: path
      required: true
      schema:
        type: string
    questionId:
      name: questionId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Question:
      type: object
      required:
        - id
        - lessonId
        - question
        - options
        - correct
        - comment
      properties:
        id:
          type: integer
        lessonId:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 5
          maxItems: 5
        correct:
          type: integer
        comment:
          type: string
    QuestionInput:
      type: object
      required:
        - question
        - options
        - correct
        - comment
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 5
          maxItems: 5
        correct:
          type: integer
        comment:
          type: string
    QuestionUpdate:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 5
          maxItems: 5
        correct:
          type: integer
        comment:
          type: string
...