# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PreClinik API — e-learning platform for Algerian medical students
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionUpdate {
  question?: string;
  /**
     * @minItems 5
     * @maxItems 5
     */
  options?: string[];
  correct?: number;
  comment?: string;
}